Broke <FunctionSpecField/> into smaller parts.
