+++ src/components/inputs/InputBlock.jsx
@@ -12,7 +12,7 @@ class InputBlock extends React.Component {
     ]).isRequired,
     doc: PropTypes.string,
     action: PropTypes.element,
-    children: PropTypes.element.isRequired,
+    children: PropTypes.node.isRequired,
     style: PropTypes.object,
     onChange: PropTypes.func,
   }
